PR-19 feat: add reaction on emoji click
